Renamed ts file and formatted using new settings
