+++ src/routes/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import RightClickableSample from '$lib/components/samplesImplementations/RightClickableSample.svelte';
+	import RightClickableSample from "$lib/components/samplesImplementations/RightClickableSample.svelte";
 
 	enum SidePanel {
 		Left,
